[AutoMPD] Only recreate file if it does not exists

+++ src/Plugins/AutoMPD.vala
@@ -615,25 +615,28 @@ public class Gmpc.Plugins.AutoMPD:
 					config_dir,
 					"autostart",
 					"%s.desktop".printf(auto_mpd_id));
-			// Write out the file:
-			FileStream? autostartfp = FileStream.open(autostart_path, "w");
-			if(autostartfp != null)
+			if(!GLib.FileUtils.test(autostart_path, GLib.FileTest.EXISTS|GLib.FileTest.IS_REGULAR))
 			{
-				var c_dir = GLib.Environment.get_user_cache_dir();
-				var full_path = GLib.Path.build_filename(c_dir,"gmpc", auto_mpd_id, "mpd.conf");
+				// Write out the file:
+				FileStream? autostartfp = FileStream.open(autostart_path, "w");
+				if(autostartfp != null)
+				{
+					var c_dir = GLib.Environment.get_user_cache_dir();
+					var full_path = GLib.Path.build_filename(c_dir,"gmpc", auto_mpd_id, "mpd.conf");
 
-				autostartfp.puts("[Desktop Entry]\n");
-				autostartfp.puts("Name=MPD\n");
-				autostartfp.puts("GenericName=Music Player Daemon (GMPC session)\n");
-				autostartfp.puts("Comment=The best way to playback music\n");
-				autostartfp.puts("Exec=mpd \"%s\"\n".printf(full_path));
-				autostartfp.puts("Terminal=false\n");
-				autostartfp.puts("Type=Application\n");
-				autostartfp.puts("Categories=Autio;Network\n");
-				autostartfp.puts("StartupNotify=false");
-			}else{
-				Gmpc.Messages.show(_("Auto MPD failed to create autostart file."),
-						Gmpc.Messages.Level.WARNING); 
+					autostartfp.puts("[Desktop Entry]\n");
+					autostartfp.puts("Name=MPD\n");
+					autostartfp.puts("GenericName=Music Player Daemon (GMPC session)\n");
+					autostartfp.puts("Comment=The best way to playback music\n");
+					autostartfp.puts("Exec=mpd \"%s\"\n".printf(full_path));
+					autostartfp.puts("Terminal=false\n");
+					autostartfp.puts("Type=Application\n");
+					autostartfp.puts("Categories=Autio;Network\n");
+					autostartfp.puts("StartupNotify=false");
+				}else{
+					Gmpc.Messages.show(_("Auto MPD failed to create autostart file."),
+							Gmpc.Messages.Level.WARNING); 
+				}
 			}
 		}
 		else
